GH#8: Work around for Field.__hash__ for django < 3.2
In Django < 3.2, instances of fields that appear on multiple models via
inheritance do not have unique hashes from __hash__ for each model they
eventually appear on. This causes the dictionary value in
previous_storages to be overwritten when the same field is used on
multiple models via inheritance as __hash__ is the same.